Add : Codechef ( Aug 17 challenge)
You have a perfect binary tree of height N. That is, it contains 2N+1 - 1 nodes.

A path from the root to a leaf can be represented by an integer X which lies between 0 and 2N - 1 (both inclusive) as follows:

Consider the N length binary representation of X. Start reading from the left, and at each bit, 0 denotes going to the left child, and 1 denotes going to the right child. For eg. if N = 2, and X = 0 = (00)2, it would correspond to the path which goes to the leftmost leaf. X = 3 = (11)2 would correspond to the path from root to the rightmost leaf.

At the beginning, X = 0, and you have visited only the root till now. Then, you get Q queries. A query can be of two types:

Type 1: "! C": This means that X has to be changed to X = ( X + 2C ) mod 2N, and then you start from the root, and walk down the route represented by the new X.
Type 2: "?": This means that you should print the number of nodes that you have visited at least once.